check 01

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: kkusunok <kkusunok@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/18 16:09:29 by kkusunok          #+#    #+#             */
-/*   Updated: 2024/05/21 15:19:30 by kkusunok         ###   ########.fr       */
+/*   Updated: 2024/06/04 10:39:51 by kkusunok         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int	ft_memcmp(const void *s1, const void *s2, size_t n)
 	while (i < n)
 	{
 		if (p1[i] != p2[i])
-			return (p1[i] - p2[i]);
+			return ((int)(p1[i] - p2[i]));
 		i++;
 	}
 	return (0);
